requirements.txt
dataclasses==0.6
Django==2.1
django-admin-tools==0.8.1
django-app-settings==0.3.0
django-braces==1.13.0
django-pkgconf==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
django-suit==0.2.25
django-suit-dashboard==2.0.6
djongo==1.2.30
nltk==3.3
Pillow==5.2.0
psutil==5.4.7
pymongo==3.7.1
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
